+++ Assets/Scripts/CombatSystem/UI/UIWinPanel.cs
@@ -13,10 +13,10 @@ namespace CombatSystem
         [SerializeField] private Button doNothingButton;
 
         public UnityEvent Evt_BtnOnClickUniversal = new UnityEvent();
-        public UnityEvent<int> Evt_OnGetLightClicked = new UnityEvent<int>();
+        public UnityEvent<float> Evt_OnGetLightClicked = new UnityEvent<float>();
         public UnityEvent<BuffType> Evt_OnGetBuffClicked = new UnityEvent<BuffType>();
 
-        private int lightAmount;
+        private float lightAmount;
         private BuffType newBuff;
 
         // Start is called before the first frame update
@@ -40,12 +40,12 @@ namespace CombatSystem
             Evt_BtnOnClickUniversal.Invoke();
         }
 
-        public void SetUpWinPanel(int light, BuffType enemyBuff, string currentBuffDescription, string newBuffDescription)
+        public void SetUpWinPanel(float light, BuffType enemyBuff, string currentBuffDescription, string newBuffDescription)
         {
             lightAmount = light;
             newBuff = enemyBuff;
 
-            getLightButton.GetComponent<UIHoverRevealer>().SetHoverText("Get " + lightAmount.ToString() + " amount of light from defeating this enemy");
+            getLightButton.GetComponent<UIHoverRevealer>().SetHoverText("Get " + ((int)lightAmount).ToString() + " amount of light from defeating this enemy");
             getBuffButton.GetComponent<UIHoverRevealer>().SetHoverText("<color=\"red\">Change from:</color>" + currentBuffDescription + "\n" + "<color=\"red\">To:</color>" + newBuffDescription);
         }
     }
